Updated programs to read in/output zipped files
Program now reads in a zipped file in the same directory as the program and searches for the FORC, MsT and Thellier files. Outputs a zipped file called processed_data.zip containing the FORC diagram and processed Thellier plots.

Updated example file names to tidy up names and make sample name in file match the title.